+++ src/sites/replicates/models.rs
@@ -11,7 +11,7 @@ pub struct SiteReplicate {
     pub name: String,
     pub sample_type: String,
     pub sampling_date: NaiveDate,
-    pub sample_depth_cm: Option<i32>,
+    pub sample_depth_cm: Option<f64>,
     pub snow_depth_cm: Option<i32>,
     pub air_temperature_celsius: Option<f64>,
     pub snow_temperature_celsius: Option<f64>,
